+++ hack/.golint_failures
@@ -344,7 +344,6 @@ pkg/securitycontext
 pkg/ssh
 pkg/util/bandwidth
 pkg/util/config
-pkg/util/configz
 pkg/util/dbus
 pkg/util/ebtables
 pkg/util/env
@@ -578,6 +577,7 @@ staging/src/k8s.io/apiserver/pkg/storage/testing
 staging/src/k8s.io/apiserver/pkg/storage/tests
 staging/src/k8s.io/apiserver/pkg/storage/value
 staging/src/k8s.io/apiserver/pkg/storage/value/encrypt/envelope/v1beta1
+staging/src/k8s.io/apiserver/pkg/util/configz
 staging/src/k8s.io/apiserver/pkg/util/feature
 staging/src/k8s.io/apiserver/pkg/util/flag
 staging/src/k8s.io/apiserver/pkg/util/proxy
